# Avoid buffer copies in `RedisStateMachine`