--- a/src/test/java/io/lettuce/core/protocol/RedisStateMachineResp2UnitTests.java
+++ b/src/test/java/io/lettuce/core/protocol/RedisStateMachineResp2UnitTests.java
@@ -74,7 +74,7 @@ class RedisStateMachineResp2UnitTests {
     @BeforeEach
     final void createStateMachine() {
         output = new StatusOutput<>(codec);
-        rsm = new RedisStateMachine(ByteBufAllocator.DEFAULT);
+        rsm = new RedisStateMachine();
     }
 
     @AfterEach