--- a/src/test/java/io/lettuce/core/protocol/RedisStateMachineResp3UnitTests.java
+++ b/src/test/java/io/lettuce/core/protocol/RedisStateMachineResp3UnitTests.java
@@ -74,7 +74,7 @@ class RedisStateMachineResp3UnitTests {
     @BeforeEach
     final void createStateMachine() {
         output = new StatusOutput<>(codec);
-        rsm = new RedisStateMachine(ByteBufAllocator.DEFAULT);
+        rsm = new RedisStateMachine();
         rsm.setProtocolVersion(ProtocolVersion.RESP3);
     }