--- a/src/test/jmh/io/lettuce/core/protocol/RedisStateMachineBenchmark.java
+++ b/src/test/jmh/io/lettuce/core/protocol/RedisStateMachineBenchmark.java
@@ -60,7 +60,7 @@ public class RedisStateMachineBenchmark {
 
     private ByteBuf masterBuffer;
 
-    private final RedisStateMachine stateMachine = new RedisStateMachine(ByteBufAllocator.DEFAULT);
+    private final RedisStateMachine stateMachine = new RedisStateMachine();
     private final byte[] payload = ("*3\r\n" + //
             "$4\r\n" + //
             "LLEN\r\n" + //