<!--
Thank you for proposing a pull request. This template will guide you through the essential steps necessary for a pull request.
-->
Make sure that:

- [x] You have read the [contribution guidelines](https://github.com/lettuce-io/lettuce-core/blob/main/.github/CONTRIBUTING.md).
- [x] You have created a feature request first to discuss your contribution intent. Please reference the feature request ticket number in the pull request.
- [x] You use the code formatters provided [here](https://github.com/lettuce-io/lettuce-core/blob/main/formatting.xml) and have them applied to your changes. Don’t submit any formatting related changes.
- [x] You submit test cases (unit or integration tests) that back your changes.

 <!--
Great! Live long and prosper.
-->

see also #2173

Here are some performance results

<details>

db.PIPELINE GET.requests - 10000
db.PIPELINE GET.size - 128
db.clients - 1
db.loop - 5
Loaded keys - 1918458
Connections Formed
PIPELINE GET Performance
Total Req   - 1280000
P50 (ms)    - 2.35
Avg (ms)    - 2.6
P90 (ms)    - 3.28
P95 (ms)    - 3.81
P99 (ms)    - 7.24
OPS         - 49167
Time Taken (s) for loop 0 is - 26.95
PIPELINE GET Performance
Total Req   - 1280000
P50 (ms)    - 2.3
Avg (ms)    - 2.84
P90 (ms)    - 3.12
P95 (ms)    - 3.56
P99 (ms)    - 6.65
OPS         - 45027
Time Taken (s) for loop 1 is - 29.26
PIPELINE GET Performance
Total Req   - 1280000
P50 (ms)    - 2.33
Avg (ms)    - 2.52
P90 (ms)    - 3.16
P95 (ms)    - 3.55
P99 (ms)    - 6.01
OPS         - 50826
Time Taken (s) for loop 2 is - 25.97
PIPELINE GET Performance
Total Req   - 1280000
P50 (ms)    - 2.32
Avg (ms)    - 2.49
P90 (ms)    - 3.13
P95 (ms)    - 3.47
P99 (ms)    - 5.8
OPS         - 51426
Time Taken (s) for loop 3 is - 25.65
PIPELINE GET Performance
Total Req   - 1280000
P50 (ms)    - 2.34
Avg (ms)    - 2.54
P90 (ms)    - 3.18
P95 (ms)    - 3.65
P99 (ms)    - 6.55
OPS         - 50337
Time Taken (s) for loop 4 is - 26.2

db.PIPELINE GET.requests - 10000
db.PIPELINE GET.size - 128
db.clients - 1
db.loop - 5
Loaded keys - 1918458
Connections Formed
PIPELINE GET Performance
Total Req   - 1280000
P50 (ms)    - 1.81
Avg (ms)    - 2.08
P90 (ms)    - 2.55
P95 (ms)    - 3.02
P99 (ms)    - 6.02
OPS         - 61661
Time Taken (s) for loop 0 is - 21.69
PIPELINE GET Performance
Total Req   - 1280000
P50 (ms)    - 1.76
Avg (ms)    - 1.93
P90 (ms)    - 2.24
P95 (ms)    - 2.56
P99 (ms)    - 4.48
OPS         - 66164
Time Taken (s) for loop 1 is - 20.18
PIPELINE GET Performance
Total Req   - 1280000
P50 (ms)    - 1.77
Avg (ms)    - 1.96
P90 (ms)    - 2.28
P95 (ms)    - 2.69
P99 (ms)    - 5.2
OPS         - 65170
Time Taken (s) for loop 2 is - 20.44
PIPELINE GET Performance
Total Req   - 1280000
P50 (ms)    - 1.75
Avg (ms)    - 1.93
P90 (ms)    - 2.22
P95 (ms)    - 2.57
P99 (ms)    - 5.09
OPS         - 66303
Time Taken (s) for loop 3 is - 20.07
PIPELINE GET Performance
Total Req   - 1280000
P50 (ms)    - 1.77
Avg (ms)    - 1.94
P90 (ms)    - 2.25
P95 (ms)    - 2.61
P99 (ms)    - 4.81
OPS         - 65856
Time Taken (s) for loop 4 is - 20.19
</details>