@mp911de 2022-10-07
Thanks a lot. Care to run the JMH benchmark `RedisStateMachineBenchmark` before and after the change? This is an interesting change that we would consider for the next minor release.