@jeffreye 2022-10-10
After
```
# JMH version: 1.21
# VM version: JDK 1.8.0_332, OpenJDK 64-Bit Server VM, 25.332-b09
# VM options: -javaagent:/Applications/IntelliJ IDEA.app/Contents/lib/idea_rt.jar=51975:/Applications/IntelliJ IDEA.app/Contents/bin -Dfile.encoding=UTF-8
# Warmup: 5 iterations, 10 s each
# Measurement: 5 iterations, 10 s each
# Timeout: 2 s per iteration, ***WARNING: The timeout might be too low!***
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Average time, time/op
# Benchmark: io.lettuce.core.protocol.RedisStateMachineBenchmark.measureDecode

# Run progress: 0.00% complete, ETA 00:01:40
# Fork: 1 of 1
# Warmup Iteration   1: 211.223 ns/op
# Warmup Iteration   2: 211.419 ns/op
# Warmup Iteration   3: 209.558 ns/op
# Warmup Iteration   4: 209.059 ns/op
# Warmup Iteration   5: 206.593 ns/op
Iteration   1: 206.514 ns/op
Iteration   2: 207.120 ns/op
Iteration   3: 206.357 ns/op
Iteration   4: 206.831 ns/op
Iteration   5: 206.158 ns/op


Result "io.lettuce.core.protocol.RedisStateMachineBenchmark.measureDecode":
  206.596 ±(99.9%) 1.474 ns/op [Average]
  (min, avg, max) = (206.158, 206.596, 207.120), stdev = 0.383
  CI (99.9%): [205.122, 208.070] (assumes normal distribution)


# Run complete. Total time: 00:01:42

REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
experiments, perform baseline and negative tests that provide experimental control, make sure
the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
Do not assume the numbers tell you what you want them to tell.

Benchmark                                 Mode  Cnt    Score   Error  Units
RedisStateMachineBenchmark.measureDecode  avgt    5  206.596 ± 1.474  ns/op

Process finished with exit code 0
```

Before

```
# JMH version: 1.21
# VM version: JDK 1.8.0_332, OpenJDK 64-Bit Server VM, 25.332-b09
# VM options: -javaagent:/Applications/IntelliJ IDEA.app/Contents/lib/idea_rt.jar=52094:/Applications/IntelliJ IDEA.app/Contents/bin -Dfile.encoding=UTF-8
# Warmup: 5 iterations, 10 s each
# Measurement: 5 iterations, 10 s each
# Timeout: 2 s per iteration, ***WARNING: The timeout might be too low!***
# Threads: 1 thread, will synchronize iterations
# Benchmark mode: Average time, time/op
# Benchmark: io.lettuce.core.protocol.RedisStateMachineBenchmark.measureDecode

# Run progress: 0.00% complete, ETA 00:01:40
# Fork: 1 of 1
# Warmup Iteration   1: 288.640 ns/op
# Warmup Iteration   2: 284.225 ns/op
# Warmup Iteration   3: 277.962 ns/op
# Warmup Iteration   4: 277.098 ns/op
# Warmup Iteration   5: 272.736 ns/op
Iteration   1: 275.259 ns/op
Iteration   2: 275.443 ns/op
Iteration   3: 279.652 ns/op
Iteration   4: 275.741 ns/op
Iteration   5: 271.997 ns/op


Result "io.lettuce.core.protocol.RedisStateMachineBenchmark.measureDecode":
  275.618 ±(99.9%) 10.467 ns/op [Average]
  (min, avg, max) = (271.997, 275.618, 279.652), stdev = 2.718
  CI (99.9%): [265.151, 286.086] (assumes normal distribution)


# Run complete. Total time: 00:01:42

REMEMBER: The numbers below are just data. To gain reusable insights, you need to follow up on
why the numbers are the way they are. Use profilers (see -prof, -lprof), design factorial
experiments, perform baseline and negative tests that provide experimental control, make sure
the benchmarking environment is safe on JVM/OS/HW level, ask for reviews from the domain experts.
Do not assume the numbers tell you what you want them to tell.

Benchmark                                 Mode  Cnt    Score    Error  Units
RedisStateMachineBenchmark.measureDecode  avgt    5  275.618 ± 10.467  ns/op

Process finished with exit code 0
```

Summary:
275.618 ± 10.467  ns/op  -> 206.596 ± 1.474  ns/op

As the payload size in this benchmark is too small, it may not reflect the reality in production environment. This change will reduce memory pressure so it will also help GC and have bigger impact to the application.