@jeffreye 2023-06-15
Hi @mp911de what's the expected date to release this change? This offers great performance improvement and we need them very much.